Move XDB helper to routing file
